cambie el tipo de la posicion de pair<int,int> a pair<float,float>, ahora se mueve bien

+++ Source/Modelo/EntidadDinamica.cpp
@@ -28,7 +28,7 @@ EntidadDinamica::EntidadDinamica(int vel,int x,int y) {
 
 }
 
-pair<int,int>* EntidadDinamica::getPosition(){
+pair<float,float>* EntidadDinamica::getPosition(){
 	return &this->position;
 }
